Read config from env variables
